Split browse_helper.rb into two modules due to rubocop ModuleLength
