+++ .rubocop_todo.yml
@@ -17,7 +17,6 @@ Lint/AssignmentInCondition:
     - 'app/controllers/users_controller.rb'
     - 'app/controllers/user_preferences_controller.rb'
     - 'app/helpers/application_helper.rb'
-    - 'app/helpers/browse_helper.rb'
     - 'app/models/client_application.rb'
     - 'app/mailers/notifier.rb'
     - 'lib/nominatim.rb'
